Move logic to forms.js
