+++ controllers/apply/form-router/views/expiry-warnings.njk
@@ -1,7 +1,7 @@
 {% from "components/modal/macro.njk" import modal %}
 
 {#
-    The IDs of these modals are referenced in session.js (client-side file)
+    The IDs of these modals are referenced in forms.js
     and must be changed there if changed here.
 #}
